feat: Only show contact button when contactable
